requirements.txt
# requirements.txt
# Application Framework
fastapi
uvicorn[standard] # Use standard for better performance

# LLM & AI Frameworks
langchain
langchain-community
langchain-core
langchain-ollama
langchain-google-genai
google-genai
langgraph

# RAG & Vector Store
faiss-cpu
scikit-learn

# Utilities
pyyaml
pydantic
python-dotenv
sse-starlette
grandalf # For langgraph visualization